Task 1 - Article Generator Chatbot/requirements.txt
streamlit==1.41.1
ollama==0.4.4
python-dotenv==1.0.1
transformers==4.47.1
huggingface-hub==0.27.0
langchain-ollama==0.2.2
google-generativeai==0.8.3
torch==1.12.1 